refactor: moved Money and Portfolio classes their own Python files
