+++ py/money.py
@@ -1,0 +1,21 @@
+# class Dollar:
+#     def __init__(self, amount):
+#         self.amount = amount
+
+#     def times(self, multiplier):
+#         return Dollar(self.amount * multiplier)
+
+
+class Money:
+    def __init__(self, amount, currency):
+        self.amount = amount
+        self.currency = currency
+
+    def times(self, multiplier):
+        return Money(self.amount * multiplier, self.currency)
+
+    def divide(self, divisor):
+        return Money(self.amount / divisor, self.currency)
+
+    def __eq__(self, other):
+        return self.amount == other.amount and self.currency == other.currency
