+++ py/test_money.py
@@ -1,52 +1,18 @@
-import functools
-import operator
+
 import unittest
 
-
-class Dollar:
-    def __init__(self, amount):
-        self.amount = amount
-
-    def times(self, multiplier):
-        return Dollar(self.amount * multiplier)
-
-
-class Money:
-    def __init__(self, amount, currency):
-        self.amount = amount
-        self.currency = currency
-
-    def times(self, multiplier):
-        return Money(self.amount * multiplier, self.currency)
-
-    def divide(self, divisor):
-        return Money(self.amount / divisor, self.currency)
-
-    def __eq__(self, other):
-        return self.amount == other.amount and self.currency == other.currency
-
-
-class Portfolio:
-    def __init__(self):
-        self.moneys = []
-
-    def add(self, *moneys):
-        self.moneys.extend(moneys)
-
-    def evaluate(self, currency):
-        total = functools.reduce(
-            operator.add, map(lambda m: m.amount, self.moneys), 0)
-        return Money(total, currency)
+from money import Money
+from portfolio import Portfolio
 
 
 class TestMoney(unittest.TestCase):
 
-    def testMultiplication(self):
-        fiver = Dollar(5)
-        tenner = fiver.times(2)
-        self.assertEqual(10, tenner.amount)
+    # def testMultiplication(self):
+    #     fiver = Dollar(5)
+    #     tenner = fiver.times(2)
+    #     self.assertEqual(10, tenner.amount)
 
-    def testMultiplicationInEuros(self):
+    def testMultiplication(self):
         tenEuros = Money(10, "EUR")
         twentyEuros = tenEuros.times(2)
         self.assertEqual(20, twentyEuros.amount)
@@ -67,10 +33,10 @@ class TestMoney(unittest.TestCase):
         self.assertEqual(expectedMoneyAfterDivision.currency,
                          actualMoneyAfterDivision.currency)
 
-    def testMultiplicationInDollars(self):
-        fiveDollars = Money(5, "USD")
-        tenDollars = Money(10, "USD")
-        self.assertEqual(tenDollars, fiveDollars.times(2))
+    # def testMultiplicationInDollars(self):
+    #     fiveDollars = Money(5, "USD")
+    #     tenDollars = Money(10, "USD")
+    #     self.assertEqual(tenDollars, fiveDollars.times(2))
 
     def testAddition(self):
         fiveDollars = Money(5, "USD")
